# engine/contract/openapi.json
{
  "components": {
    "schemas": {
...
      "EndpointRequest": {
        "description": "One slot's HTTP endpoint. Omitted fields keep their current value.\n\n``api_key`` is the reason for that rule rather than a plain replace: the\nclient never receives the key back (see\n:func:`~wrenote.core.openai_compat.endpoint_status`), so it cannot send it\nagain, and a form save must not wipe it. Passing ``\"\"`` explicitly clears\nit — which is how the UI's \"forget the key\" works.",
        "properties": {
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "api_key_env": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key Env"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "timeout_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout S"
          }
        },
        "required": [
          "kind"
        ],
        "title": "EndpointRequest",
        "type": "object"
      },
      "FeaturesRequest": {
        "description": "The optional features to switch on or off; omitted slots are left alone.",
        "properties": {
          "chat": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat"
          },
          "speaker": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speaker"
          },
          "translator": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Translator"
          }
        },
        "title": "FeaturesRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "wrenote__api__models__SelectRequest": {
        "properties": {
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "model"
        ],
        "title": "SelectRequest",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Wrenote Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/models/endpoint": {
      "post": {
        "description": "Point a slot at a model served over HTTP, persisting to the user config.\n\nSame \"applies now / next session\" split as :func:`models_select`: chat is\nheld by :class:`ModelManager` and swapped here; the translator is built per\nsession. Switching ``active`` off restores the slot's catalogue model\nrather than leaving it on a backend it is no longer meant to use.",
        "operationId": "models_endpoint_v1_models_endpoint_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EndpointRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Models Endpoint V1 Models Endpoint Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Models Endpoint"
      }
    },
    "/v1/models/endpoint/test": {
      "post": {
        "description": "Ask the configured endpoint one tiny question, and report what happened.\n\nThis is the probe ``load()`` deliberately doesn't do (a health check on\nevery start-up costs a round trip, and some shims serve\n``/chat/completions`` and nothing else). Here it is worth it: the moment\nsomeone presses Save is exactly when a typo in a URL or a stale key should\nsurface, rather than three days later mid-meeting.\n\nTests what is *saved*, not what is typed — the key may only exist in the\nconfig — so the client saves first and tests second.",
        "operationId": "models_endpoint_test_v1_models_endpoint_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EndpointRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Models Endpoint Test V1 Models Endpoint Test Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Models Endpoint Test"
      }
    },
    "/v1/models/features": {
      "post": {
        "description": "Switch optional features on or off, persisting to the user config.\n\nSwitching one on does not download anything: the caller follows with\n``POST /models/download``, which now sees the slot in ``required_models``.\nSwitching one off keeps the files — deleting a 2.5 GB model because a\ntoggle moved is not a thing to do silently — but drops the running\nbackend, so the memory goes back.",
        "operationId": "models_features_v1_models_features_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeaturesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Models Features V1 Models Features Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Models Features"
      }
    },
    "/v1/models/select": {
      "post": {
        "description": "Choose the model for one kind, persisting it to the user config.\n\nHow soon it applies differs by kind, and the response says which: STT and\nthe translator are constructed per WebSocket session, so the next session\npicks the new model up; chat and the diarization speaker are held by\n:class:`ModelManager`, so they are swapped here and now. Neither needs a\nrestart — claiming otherwise would train people to restart for nothing.",
        "operationId": "models_select_v1_models_select_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/wrenote__api__models__SelectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Models Select V1 Models Select Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Models Select"
      }
    },
...
    "/v1/models/{model_id}": {
      "delete": {
        "description": "Delete a catalogue model's files from ``models.dir``.\n\nFor testing what the app does when a model is absent — the first-run\nwizard, the download progress, a slot with nothing to run — which\notherwise means finding the files by hand. Recoverable by definition:\nthe catalogue knows where every file came from, so the next download\nfetches it again.\n\nA model that is currently loaded is dropped from the manager first;\notherwise Windows would refuse to unlink the file under it.",
        "operationId": "models_delete_v1_models__model_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "title": "Model Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Models Delete V1 Models  Model Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Models Delete"
      }
    },
...
  }
}